+++ changelog.md
@@ -5,6 +5,7 @@
 - new: sources can contain "/" (and will use subdirectories)
 - gui: new Properties page for stories
 - gui: rename stories, change author
+- gui: allow "all" and "listing" modes for sources and authors
 - tui: now working well enough to be considered stable
 - remote: fix setSourceCover (was not seen by client)
 - remote: can now import local files into a remote library